bug fix in internal evaluation

+++ Internal Evaluation/internal_evaluation.py
@@ -132,7 +132,7 @@ sys.stdout.write("\r%d/%d\n" % (a + 1, test_cases))
 
 if net_type=="snn":
     file = open("../Results/SNN_results", "a")
-    parameter_count = str(978*d*k + 0.5*k*k*d*(d-1))
+    parameter_count = str(int(978*d*k + 0.5*k*k*d*(d-1)))
 
     outputs = [parameter_count, str(d), str(k), str(embedding_length), 'MOD'+embedding_name[3:], embedding_name]
